Se realizan ajuste en tipos de datos y textos para mostrar en pantalla

+++ src/main/java/models/Administrador.java
@@ -2,30 +2,30 @@ package models;
 
 public class Administrador extends Persona {
 
-    private Float gananciaEnero;
+    private Integer gananciaEnero;
 
 
     public Administrador() {
     }
 
-    public Administrador(Float gananciaEnero) {
+    public Administrador(Integer gananciaEnero) {
         this.gananciaEnero = gananciaEnero;
     }
 
-    public Float getGananciaEnero() {
+    public Integer getGananciaEnero() {
         return gananciaEnero;
     }
 
-    public void setGananciaEnero(Float gananciaEnero) {
+    public void setGananciaEnero(Integer gananciaEnero) {
         this.gananciaEnero = gananciaEnero;
     }
 
-    public Administrador(String id, Float gananciaEnero) {
+    public Administrador(String id, Integer gananciaEnero) {
         super(id);
         this.gananciaEnero = gananciaEnero;
     }
 
-    public Administrador(String id, String nombre, String apellido, Float gananciaEnero) {
+    public Administrador(String id, String nombre, String apellido, Integer gananciaEnero) {
         super(id, nombre, apellido);
         this.gananciaEnero = gananciaEnero;
     }
